Remove INTERNAL_IPS from docs & example
This used to be required, but AFAICT it no longer is.

(As I understand it, it is from the days when the `{% tailwind_css %}` tag had
auto-reloading by default, but only in the case when your IP address matched
`INTERNAL_IPS`. This has since been made explicit with a separate tag and a
TAILWIND_DEV_MODE setting.)

I've grepped on 'debug' and 'INTERNAL_IPS' to make sure.