Revert "Drops Python 3.9 support"
This reverts commit 4d7a815298fa

See https://github.com/timonweb/django-tailwind/issues/210

Python 3.9 has a half year of life in it left.
Support for it was dropped but it's not clear to me whether there was an actual
direct cause for this or whether it was simply "let's get rid of old(ish)
stuff".

Since this was never answered I'm just going to assume "no reason" and revert
this for now.